feature: impl auth mvp

+++ src/user/routes.rs
@@ -1,50 +1,60 @@
 use crate::api_error::ApiError;
+use crate::auth::{require_owner, AuthUser};
 use crate::response::Response;
-use crate::user::{CreateUserDto, UpdateUserDto, User};
-use actix_web::{delete, get, post, put, web, HttpResponse};
+use crate::user::{PublicUser, UpdateUserDto, User};
+use actix_web::{delete, get, put, web, HttpResponse};
 use serde_json::json;
 use uuid::Uuid;
 
 #[get("/users")]
 async fn find_all() -> Result<HttpResponse, ApiError> {
     let users = User::find_all()?;
+    let users: Vec<PublicUser> = users
+        .into_iter()
+        .map(|u| PublicUser::from(u))
+        .collect::<Vec<PublicUser>>();
 
     Ok(HttpResponse::Ok().json(Response::from(users)))
 }
 
-#[get("/users/{handle}")]
+#[get("/users/{id}")]
 async fn find(id: web::Path<Uuid>) -> Result<HttpResponse, ApiError> {
     let user_id = id.into_inner();
     let user = User::find(user_id)?;
+    let user = PublicUser::from(user);
 
     Ok(HttpResponse::Ok().json(user))
 }
 
-// TODO: Auth routes should be in a separate module
-#[post("/auth/register")]
-async fn register(user: web::Json<CreateUserDto>) -> Result<HttpResponse, ApiError> {
-    let user = User::create(user.into_inner())?;
-
-    Ok(HttpResponse::Ok().json(user))
-}
-
-#[put("/users/{handle}")]
+#[put("/users/{id}")]
 async fn update(
     id: web::Path<Uuid>,
     user: web::Json<UpdateUserDto>,
+    auth: AuthUser,
 ) -> Result<HttpResponse, ApiError> {
     let user_id = id.into_inner();
+    require_owner(user_id, auth)?;
     let user = User::update(user_id, user.into_inner())?;
+    let user = PublicUser::from(user);
 
     Ok(HttpResponse::Ok().json(user))
 }
 
-#[delete("/users/{handle}")]
-async fn delete(id: web::Path<Uuid>) -> Result<HttpResponse, ApiError> {
+#[delete("/users/{id}")]
+async fn delete(id: web::Path<Uuid>, auth: AuthUser) -> Result<HttpResponse, ApiError> {
     let user_id = id.into_inner();
+    require_owner(user_id, auth)?;
     let num_deleted = User::delete(user_id)?;
 
     Ok(HttpResponse::Ok().json(json!({ "deleted": num_deleted })))
+}
+
+#[get("/profile")]
+async fn profile(auth: AuthUser) -> Result<HttpResponse, ApiError> {
+    let user = User::find(auth.id)?;
+    let user = PublicUser::from(user);
+
+    Ok(HttpResponse::Ok().json(user))
 }
 
 pub fn init_routes(cfg: &mut web::ServiceConfig) {
@@ -52,7 +62,5 @@ pub fn init_routes(cfg: &mut web::ServiceConfig) {
     cfg.service(find);
     cfg.service(update);
     cfg.service(delete);
-
-    // TODO: Auth routes should be in a separate module
-    cfg.service(register);
+    cfg.service(profile);
 }
